Rename register -> registration.py
