+++ corvid/main/forms/login.py
@@ -1,0 +1,6 @@
+from django import forms
+
+
+class LoginForm(forms.Form):
+    username = forms.CharField(max_length=30)
+    password = forms.CharField(widget=forms.PasswordInput)
